# Major ZIR size optimizations & small cleanups in Sema