Two things here:
* A couple of optimizations in `AstGen`, most notably elision of a bunch of unnecessary `block` instructions. This results in a big drop in ZIR bytes.
* Cleanups to the main loop of `Sema` to improve simplicitly and readability.

According to my quick testing, these changes have minimal performance impact, but it's a benefit where it exists, e.g.:
```
Benchmark 1 (30 runs): /home/mlugg/zig/master/stage4-fast/bin/zig test -fno-emit-bin test/behavior.zig
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           688ms ± 9.42ms     669ms …  711ms          0 ( 0%)        0%
  peak_rss            104MB ±  160KB     104MB …  104MB          0 ( 0%)        0%
  cpu_cycles         2.28G  ± 32.5M     2.22G  … 2.37G           0 ( 0%)        0%
  instructions       2.80G  ± 6.95K     2.80G  … 2.80G           0 ( 0%)        0%
  cache_references    288M  ± 2.05M      283M  …  292M           0 ( 0%)        0%
  cache_misses       42.2M  ±  927K     40.6M  … 44.4M           0 ( 0%)        0%
  branch_misses      14.5M  ± 86.9K     14.3M  … 14.8M           1 ( 3%)        0%
Benchmark 2 (30 runs): /home/mlugg/zig/dbg-var-blocks/stage4-fast/bin/zig test -fno-emit-bin test/behavior.zig
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           669ms ± 11.6ms     649ms …  694ms          0 ( 0%)        ⚡-  2.8% ±  0.8%
  peak_rss            103MB ±  173KB     102MB …  103MB          0 ( 0%)        ⚡-  1.5% ±  0.1%
  cpu_cycles         2.23G  ± 33.8M     2.19G  … 2.32G           0 ( 0%)        ⚡-  2.1% ±  0.8%
  instructions       2.81G  ± 11.4K     2.81G  … 2.81G           0 ( 0%)          +  0.2% ±  0.0%
  cache_references    275M  ± 2.39M      271M  …  283M           1 ( 3%)        ⚡-  4.4% ±  0.4%
  cache_misses       40.3M  ± 1.11M     38.1M  … 43.4M           1 ( 3%)        ⚡-  4.6% ±  1.3%
  branch_misses      13.9M  ±  100K     13.8M  … 14.2M           1 ( 3%)        ⚡-  4.0% ±  0.3%
```